+++ combinatoric.hh
@@ -15,7 +15,7 @@ uint choose(uint n, uint k);
 long double ldchoose(uint n, uint k);
 
 // greatest common divisor via the Euclidean algorithm
-long long gcd64(unsigned long long n1, unsigned long long n2);
+long gcd64(unsigned long n1, unsigned long n2);
 
 // greatest common divisor via the Euclidean algorithm
 uint gcd(uint n1, uint n2);
